# openApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PaymentGateway",
    "version": "v1"
  },
  "paths": {
    "/v1/Payment": {
      "post": {
        "tags": [
          "Payment"
        ],
        "operationId": "RequestPayment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SucessfulPaymentRequestResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SucessfulPaymentRequestResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SucessfulPaymentRequestResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FailedPaymentRequestResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedPaymentRequestResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedPaymentRequestResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/Payment/{trackingId}": {
      "get": {
        "tags": [
          "Payment"
        ],
        "operationId": "GetPaymentRequest",
        "parameters": [
          {
            "name": "trackingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MerchantPaymentRequestFailedToProcess"
                    },
                    {
                      "$ref": "#/components/schemas/MerchantPaymentRequestSuccessfullyProcessed"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MerchantPaymentRequestFailedToProcess"
                    },
                    {
                      "$ref": "#/components/schemas/MerchantPaymentRequestSuccessfullyProcessed"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MerchantPaymentRequestFailedToProcess"
                    },
                    {
                      "$ref": "#/components/schemas/MerchantPaymentRequestSuccessfullyProcessed"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BankTransactionResult": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/BankTransactionResultStatus"
          },
          "bankTransactionId": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BankTransactionResultStatus": {
        "enum": [
          "NotFound",
          "Rejected",
          "Accepted"
        ],
        "type": "string"
      },
      "CardPaymentSource": {
        "type": "object",
        "properties": {
          "issuer": {
            "type": "string",
            "nullable": true
          },
          "cvv": {
            "type": "integer",
            "format": "int32"
          },
          "cardHolder": {
            "type": "string",
            "nullable": true
          },
          "cardNumber": {
            "type": "string",
            "nullable": true
          },
          "expiry": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CompletedMerchantPaymentRequestEvent": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/Event"
          }
        ],
        "properties": {
          "merchantId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "paymentRequest": {
            "$ref": "#/components/schemas/PaymentRequest"
          }
        },
        "additionalProperties": false
      },
      "Event": {
        "type": "object",
        "properties": {
          "trackingId": {
            "type": "string",
            "nullable": true
          },
          "timeStamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "FailedPaymentRequestResult": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/PaymentRequestResult"
          }
        ],
        "properties": {
          "errorType": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MerchantPaymentRequestFailedToProcess": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/CompletedMerchantPaymentRequestEvent"
          }
        ],
        "properties": {
          "errorType": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "MerchantPaymentRequestSuccessfullyProcessed": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/CompletedMerchantPaymentRequestEvent"
          }
        ],
        "properties": {
          "bankTransactionResult": {
            "$ref": "#/components/schemas/BankTransactionResult"
          },
          "status": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "PaymentRequest": {
        "type": "object",
        "properties": {
          "currency": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "source": {
            "$ref": "#/components/schemas/CardPaymentSource"
          }
        },
        "additionalProperties": false
      },
      "PaymentRequestResult": {
        "type": "object",
        "properties": {
          "trackingId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "SucessfulPaymentRequestResult": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/PaymentRequestResult"
          }
        ],
        "additionalProperties": false
      },
      "ValidationProblemDetails": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ProblemDetails"
          }
        ],
        "properties": {
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}